refactor(airflow): upgrade to airflow v2
