+++ airflow/dags/gtfs_views/METADATA.yml
@@ -18,4 +18,4 @@ default_args:
     #sla: !timedelta 'hours: 2'
 wait_for_defaults:
     timeout: 3600
-latest_only: False
+latest_only: True
