+++ airflow/plugins/operators/__init__.py
@@ -1,9 +1,11 @@
 # flake8: noqa
-from .external_table import ExternalTable
-from .once_off_external_task_sensor import OnceOffExternalTaskSensor
-from .pod_operator import PodOperator
-from .python_taskflow_operator import PythonTaskflowOperator
-from .sql_to_warehouse_operator import SqlToWarehouseOperator
-from .csv_to_warehouse_operator import CsvToWarehouseOperator
-from .python_to_warehouse_operator import PythonToWarehouseOperator
-from .sql_query_operator import SqlQueryOperator
+# TODO: after upgrading airflow, using a relative import raises the error
+# "attempted relative import with no known parent package"
+from operators.external_table import ExternalTable
+from operators.once_off_external_task_sensor import OnceOffExternalTaskSensor
+from operators.pod_operator import PodOperator
+from operators.python_taskflow_operator import PythonTaskflowOperator
+from operators.sql_to_warehouse_operator import SqlToWarehouseOperator
+from operators.csv_to_warehouse_operator import CsvToWarehouseOperator
+from operators.python_to_warehouse_operator import PythonToWarehouseOperator
+from operators.sql_query_operator import SqlQueryOperator
