+++ airflow/plugins/operators/python_to_warehouse_operator.py
@@ -1,6 +1,6 @@
 import inspect
 
-from airflow.operators import PythonOperator
+from airflow.operators.python import PythonOperator
 from calitp.config import format_table_name
 from calitp.sql import sql_patch_comments
