+++ airflow/plugins/operators/sql_query_operator.py
@@ -2,8 +2,6 @@
 # flake8: noqa
 
 from airflow.models import BaseOperator
-from airflow.utils.decorators import apply_defaults
-
 from calitp import get_engine
 from sqlalchemy import sql
 
@@ -12,7 +10,6 @@ class SqlQueryOperator(BaseOperator):
 
     template_fields = ("sql",)
 
-    @apply_defaults
     def __init__(self, sql, **kwargs):
         super().__init__(**kwargs)
